Little test refactor
